+++ db/migrate/20190205200827_add_role_columns_to_users.rb
@@ -1,0 +1,9 @@
+class AddRoleColumnsToUsers < ActiveRecord::Migration[6.0]
+  def change
+    add_column :users, :username, :string
+    add_column :users, :role, :string
+    add_column :users, :phone, :string
+    add_column :users, :keywords, :string
+    add_column :users, :profile_picture, :string
+  end
+end