+++ test/controllers/settings_controller_test.rb
@@ -1,0 +1,154 @@
+# frozen_string_literal: true
+
+require "test_helper"
+
+# Test user settings pages.
+class SettingsControllerTest < ActionDispatch::IntegrationTest
+  setup do
+    @regular = users(:regular)
+  end
+
+  test "should get settings" do
+    login(@regular)
+    get settings_url
+    assert_redirected_to settings_profile_url
+  end
+
+  test "should get profile for regular user" do
+    login(@regular)
+    get settings_profile_url
+    assert_response :success
+  end
+
+  test "should not get profile for public user" do
+    get settings_profile_url
+    assert_redirected_to new_user_session_url
+  end
+
+  test "should complete profile" do
+    login(@regular)
+    patch settings_complete_profile_url, params: {
+      user: {
+        role: "Staff Member",
+        phone: "555-555-5555"
+      }
+    }
+    @regular.reload
+    assert_equal "Staff Member", @regular.role
+    assert_equal "555-555-5555", @regular.phone
+    assert_equal "Profile successfully updated.", flash[:notice]
+    assert_redirected_to dashboard_url
+  end
+
+  test "should update profile" do
+    login(@regular)
+    patch settings_update_profile_url, params: {
+      user: {
+        username: "regularupdate",
+        role: "Staff Member",
+        phone: "555-555-5555"
+      }
+    }
+    @regular.reload
+    assert_equal "regularupdate", @regular.username
+    assert_equal "Staff Member", @regular.role
+    assert_equal "555-555-5555", @regular.phone
+    assert_equal "Profile successfully updated.", flash[:notice]
+    assert_redirected_to settings_profile_url
+  end
+
+  test "should update profile picture" do
+    login(@regular)
+    patch settings_update_profile_picture_url, params: {
+      user: {
+        profile_picture: fixture_file_upload(file_fixture("rails.png"))
+      }
+    }
+    @regular.reload
+    assert_equal true, @regular.profile_picture.present?
+    assert_equal "Profile picture successfully updated.", flash[:notice]
+    assert_redirected_to settings_profile_url
+  end
+
+  test "should get account" do
+    login(@regular)
+    get settings_account_url
+    assert_response :success
+  end
+
+  test "should update account" do
+    login(@regular)
+    patch settings_update_account_url, params: {
+      user: { username: "newusername" }
+    }
+    assert_equal "Account successfully updated.", flash[:notice]
+    assert_redirected_to settings_account_url
+  end
+
+  test "should update password" do
+    sign_in_as(@regular, "password")
+    patch settings_update_password_url, params: {
+      user: {
+        current_password: "password",
+        password: "newpassword",
+        password_confirmation: "newpassword"
+      }
+    }
+    assert_equal "Your password has been changed.", flash[:notice]
+    assert_redirected_to settings_account_url
+  end
+
+  test "should not update password as user with invalid current password" do
+    sign_in_as(@regular, "password")
+    patch settings_update_password_url, params: {
+      user: {
+        current_password: "invalid",
+        password: "newpassword",
+        password_confirmation: "newpassword"
+      }
+    }
+    assert_response :success
+  end
+
+  test "should not update password with new password mismatch" do
+    sign_in_as(@regular, "password")
+    patch settings_update_password_url, params: {
+      user: {
+        current_password: "password",
+        password: "newpassword",
+        password_confirmation: "mismatched"
+      }
+    }
+    assert_response :success
+  end
+
+  # test "should delete account" do
+  #   login(@regular)
+  #   assert_difference("User.current.count", -1) do
+  #     delete settings_delete_account_url
+  #   end
+  #   assert_redirected_to root_url
+  # end
+
+  test "should get email" do
+    login(@regular)
+    get settings_email_url
+    assert_response :success
+  end
+
+  test "should update email" do
+    login(@regular)
+    patch settings_update_email_url, params: { user: { email: "newemail@example.com" } }
+    @regular.reload
+
+    # These three are correct if :confirmable is enabled
+    # assert_equal "regular@example.com", @regular.email
+    # assert_equal "newemail@example.com", @regular.unconfirmed_email
+    # assert_equal I18n.t("devise.confirmations.send_instructions"), flash[:notice]
+    # Otherwise the following two lines is correct:
+    assert_equal "newemail@example.com", @regular.email
+    assert_equal "Your email has been changed.", flash[:notice]
+
+    assert_redirected_to settings_email_url
+  end
+end