+++ CHANGELOG.md
@@ -1,6 +1,8 @@
 ## 14.0.0
 
 ### Enhancements
+- **General Changes**
+  - Add settings page to allow users to update their email and password
 - **Gem Changes**
   - Update to ruby 2.6.1
   - Update to rails 6.0.0.beta1